backend/requirements.txt
asgiref
Django
psycopg2-binary
sqlparse
whitenoise
tzdata
djangorestframework
mysqlclient